mfv now runs
